latest updates. moved requests and handlers to a hook
