+++ src/pages/Chat/Chat.tsx
@@ -1,51 +1,19 @@
 /* eslint-disable react/no-children-prop */
-import {
-  useGetMessagesNotificationsQuery,
-  useGetRoomsQuery,
-} from "services/notifications/setNotificationsAPI";
-import { IUser, useGetOwnUserQuery } from "services/user/setUserAPI";
 import { Block, Container, Notification } from "pages/Chat/Chat.styles";
-import { useTranslation } from "react-i18next";
-import { IChatUser } from "services/notifications/chatTypes";
 import ChatRoomsListItem from "pages/Chat/components/ChatRoomsList/ChatRoomsListItem";
 import { NavLink, Route, Routes } from "react-router-dom";
 import ChatRoom from "pages/Chat/components/ChatRoom/ChatRoom";
 import { CHAT_ROOM_ROUTE } from "utils/consts/routeConsts";
+import useChatData from "./hooks/useChatData";
 
 function Chat() {
-  const { t } = useTranslation();
-
-  const { data: roomsList } = useGetRoomsQuery();
-
-  const { data: user } = useGetOwnUserQuery();
-
-  const { data: notificationsResponse } = useGetMessagesNotificationsQuery();
-
-  const getRoomNotifications = (roomId: number): number[] | undefined => {
-    if (notificationsResponse && notificationsResponse.notifications) {
-      return notificationsResponse.notifications
-        .filter((notification) => notification.roomId === roomId)
-        .map((notification) => notification.id);
-    }
-    return undefined;
-  };
-
-  const defineOpponentsNameAndAvatar = (
-    freelancer: IChatUser,
-    client: IChatUser,
-    currentUser: IUser
-  ): { name: string; avatar: string } => {
-    if (+currentUser.id! !== freelancer.id) {
-      return {
-        name: `${freelancer.firstName || ""} ${freelancer.lastName || ""}`,
-        avatar: `${freelancer.avatarURL || ""}`,
-      };
-    }
-    return {
-      name: `${client.firstName || ""} ${client.lastName || ""}`,
-      avatar: `${client.avatarURL || ""}`,
-    };
-  };
+  const {
+    t,
+    user,
+    roomsList,
+    getRoomNotifications,
+    defineOpponentsNameAndAvatar,
+  } = useChatData();
 
   return (
     <div>
