+++ src/pages/Notification/style.ts
@@ -6,7 +6,7 @@ import {
   LIGHT_GRAY,
 } from "utils/consts/colorConsts";
 
-export const NotificationBox = styled.div<{ isRead: boolean }>`
+const NotificationBox = styled.div<{ isRead: boolean }>`
   width: 230px;
   border-radius: 12px;
   background-color: ${({ isRead }) => (isRead ? LIGHT_GRAY : LIGHT_BLUE)};
@@ -15,6 +15,13 @@ export const NotificationBox = styled.div<{ isRead: boolean }>`
   cursor: pointer;
   box-shadow: ${BOX_SHADOW_LIGHT} 0px 1px 3px, ${BOX_SHADOW_DARK} 0px 1px 2px;
   display: flex;
+  flex-direction: column;
+  align-items: center;
+`;
+
+export const Container = styled.div`
+  display: flex;
+  align-items: center;
   flex-direction: column;
 `;
 export default NotificationBox;
